restructure processors
This Commit does the following:

 - moves processors into their own sub-packages
 - adds ability to add any typed field value into Data
