+++ plugins/ingest/src/main/java/org/elasticsearch/ingest/Pipeline.java
@@ -20,6 +20,8 @@
 
 package org.elasticsearch.ingest;
 
+import org.elasticsearch.ingest.processor.Processor;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
